Add BrownoutDetected Events - zuma sepolicy
Brownout Detection is detected during the boot sequence.  If the
previous shutdown resulted in a reboot reason that has *ocp* or *uvlo*
in it, the shutdown was due to brownout.  Mitigation Logger should have
logged the device state during the brownout.  This event metric is to
surface the logged data.

Bug: 250009365
Test: Confirm triggering of events
Ignore-AOSP-First: to detect brownout.
Change-Id: I54e354372935e339f685fc8f5541a0568053ae08
Signed-off-by: George Lee <geolee@google.com>